UPDATE
- Sell minerals mechanic -> turn this into some sort of level/ xp system too
- Some changes in the tutorial
- UI changed

+++ Assets/Res/Entities/EntityStatData.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class EntityStatData : ScriptableObject {
     public float health = 40;
     public float currentHealth = 40;
-    public float energy = 30;
-    public float currentEnergy = 30;
+    public int energy = 30;
+    public int currentEnergy = 30;
     public int speed = 1;
     public Moveset[] movesets;
